requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
numpy==1.16.5
pandas==0.25.1
requests==2.22.0
beautifulsoup4==4.8.0
ConfigParser==5.0.2
pymong==3.11.4
bson==0.5.10